requirements.txt
Pillow==10.2.0
cryptography==42.0.5
pyfiglet==1.0.2
termcolor==2.4.0

